python/requirements.txt
requests==2.31.0
pytest==7.4.0
pytest-html==4.0.0
deepdiff==6.3.1
jsonschema==4.19.0
python-dotenv==1.0.0
